# docs/openapi.yaml
...
components:
  schemas:
...
    BlockSlotRequest:
      description: 'An administrative reservation. Always exclusive: a hall closed for repairs

        is closed to everybody, whatever its booking mode says.'
      properties:
        date:
          format: date
          title: Date
          type: string
        department:
          anyOf:
          - maxLength: 120
            type: string
          - type: 'null'
          title: Department
        endTime:
          title: Endtime
          type: string
        notes:
          anyOf:
          - maxLength: 2000
            type: string
          - type: 'null'
          title: Notes
        reason:
          maxLength: 200
          minLength: 1
          title: Reason
          type: string
        startTime:
          title: Starttime
          type: string
      required:
      - reason
      - date
      - startTime
      - endTime
      title: BlockSlotRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Page_AmenitySummary_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/AmenitySummary'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[AmenitySummary]
      type: object
    Page_ApprovalRequest_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/ApprovalRequest'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[ApprovalRequest]
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  summary: Backend for the HomeBandhu residential community platform.
  title: HomeBandhu API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/amenities:
    get:
      description: 'The catalogue grid, alphabetically.


        ``pendingRequests`` and ``outstandingDues`` are derived from the approvals

        queue and the ledger rather than stored, so the two badges on a card cannot

        disagree with the tabs they link to. The seeded data stores both as constants

        and they are already wrong there.


        Inactive amenities are included: the resident screen renders them greyed out

        rather than hiding them, so filtering them away here would change what

        residents see.'
      operationId: list_amenities_api_v1_amenities_get
      parameters:
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Q
      - description: Sports | Fitness | Recreation | Events | Utility
        in: query
        name: category
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Sports | Fitness | Recreation | Events | Utility
          title: Category
      - description: Active | Inactive | All
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Active | Inactive | All
          title: Status
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: pageSize
        required: false
        schema:
          default: 24
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AmenitySummary_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List amenities
      tags:
      - amenities
    post:
      description: 'Create an amenity, with a full settings row from the defaults.


        Returns 409 when the community already has an amenity of that name -- one

        unique index, so two admins creating "Clubhouse Gym" at the same moment

        cannot both succeed.'
      operationId: create_amenity_api_v1_amenities_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAmenityRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmenityDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create an amenity
      tags:
      - amenities
  /api/v1/amenities/{amenity_id}:
    delete:
      description: 'Delete an amenity nobody has booked.


        Returns 409 once anything has been booked on it, naming the count and

        pointing at deactivation instead: the cascade would take the bookings, their

        charges and their financial events with it, including deposits residents are

        still owed.'
      operationId: delete_amenity_api_v1_amenities__amenity_id__delete
      parameters:
      - in: path
        name: amenity_id
        required: true
        schema:
          title: Amenity Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete an amenity
      tags:
      - amenities
    get:
      description: One amenity with all five settings groups the settings tab edits.
      operationId: get_amenity_api_v1_amenities__amenity_id__get
      parameters:
      - in: path
        name: amenity_id
        required: true
        schema:
          title: Amenity Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmenityDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get an amenity
      tags:
      - amenities
    patch:
      description: 'Update the catalogue fields, the settings, or both, in one transaction.


        The settings tab saves thirty fields on one click; splitting that across two

        calls would let the second fail after the first had already been written.

        A field left out is left alone.'
      operationId: update_amenity_api_v1_amenities__amenity_id__patch
      parameters:
      - in: path
        name: amenity_id
        required: true
        schema:
          title: Amenity Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAmenityRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmenityDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update an amenity
      tags:
      - amenities
  /api/v1/amenities/{amenity_id}/approvals:
    get:
      description: 'The approvals tab -- **one row per request, not per day**.


        The frontend produces one row per day and approves them one at a time, so a

        three-day request appears three times and can be approved on Monday and

        rejected on Tuesday. Here one decision covers the request, and ``dayCount``

        and ``dates`` say how much it covers.


        ``outstandingDues`` is the **flat''s** balance across its unpaid invoices. The

        frontend computes it per person; our invoices attach to the unit and carry no

        person, so it is the same label over a different number

        (DECISIONS_NEEDED E18).'
      operationId: list_approvals_api_v1_amenities__amenity_id__approvals_get
      parameters:
      - in: path
        name: amenity_id
        required: true
        schema:
          title: Amenity Id
          type: string
      - description: pending | approved | rejected | cancelled | all
        in: query
        name: status
        required: false
        schema:
          default: pending
          description: pending | approved | rejected | cancelled | all
          title: Status
          type: string
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: pageSize
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ApprovalRequest_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List booking requests awaiting a decision
      tags:
      - amenities
  /api/v1/amenities/{amenity_id}/blocks:
    post:
      description: 'Reserve a slot for maintenance or an association event.


        Always exclusive, whatever the amenity''s booking mode says: a hall closed for

        repairs is closed to everybody. Carries no flat and no charges.'
      operationId: block_slot_api_v1_amenities__amenity_id__blocks_post
      parameters:
      - in: path
        name: amenity_id
        required: true
        schema:
          title: Amenity Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockSlotRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Block a slot administratively
      tags:
      - amenities
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Deployed
  url: https://{host}
  variables:
    host:
      default: api.homebandhu.example
      description: API host
...